Refactored entity modules. Finished EagleEye API class.
