+++ tests/fixtures/carson.live/carson_me.json
@@ -19,7 +19,7 @@
             "buildingName": "122 Fake St",
             "name": "Camera 1",
             "provider": "eagle_eye",
-            "externalId": "142c7001",
+            "externalId": "c0",
             "staffOnly": false,
             "defaultInBuilding": true,
             "liveViewId": "5511f001",
@@ -36,7 +36,7 @@
             "buildingName": "122 Fake St",
             "name": "Camera 2",
             "provider": "eagle_eye",
-            "externalId": "5511f001",
+            "externalId": "c1",
             "staffOnly": false,
             "defaultInBuilding": false,
             "liveViewId": "100c7a1d",
