+++ tests/fixtures/carson.live/carson_me_update.json
@@ -19,7 +19,7 @@
             "buildingName": "122 Fake St_update",
             "name": "Camera 1_update",
             "provider": "eagle_eye",
-            "externalId": "142c7001_update",
+            "externalId": "c0",
             "staffOnly": false,
             "defaultInBuilding": true,
             "liveViewId": "5511f001_update",
@@ -36,7 +36,7 @@
             "buildingName": "122 Fake St_new",
             "name": "Camera 3_new",
             "provider": "eagle_eye",
-            "externalId": "5511f001_new",
+            "externalId": "c2",
             "staffOnly": false,
             "defaultInBuilding": false,
             "liveViewId": "100c7a1d_new",
@@ -51,7 +51,7 @@
             "buildingName": "122 Fake St_new",
             "name": "Camera 4_new",
             "provider": "eagle_eye",
-            "externalId": "5511f001_new",
+            "externalId": "c3",
             "staffOnly": false,
             "defaultInBuilding": false,
             "liveViewId": "100c7a1d_new",
