+++ tests/fixtures/eagleeyenetworks.com/device_camera_update.json
@@ -3,7 +3,7 @@
     "bridges_key": "bridges_value"
   },
   "camera_info_status_code": 200,
-  "name": "Camera",
+  "name": "Updated_Camera_Name",
   "settings": {
     "settings_key": "settings_value"
   },
