+++ tests/fixtures/eagleeyenetworks.com/device_list.json
@@ -1,0 +1,433 @@
+[
+  [
+    "",
+    "c0",
+    "",
+    "camera",
+    [
+      [
+        "",
+        ""
+      ]
+    ],
+    "",
+    "",
+    [
+      ""
+    ],
+    "",
+    null,
+    0,
+    "",
+    0,
+    0,
+    "",
+    0,
+    "",
+    false,
+    null,
+    null,
+    [
+      null,
+      null,
+      null,
+      null,
+      "",
+      null,
+      ""
+    ],
+    null,
+    null,
+    0,
+    [],
+    0,
+    {},
+    ""
+  ],
+  [
+    "",
+    "c1",
+    "",
+    "camera",
+    [
+      [
+        "",
+        ""
+      ]
+    ],
+    "",
+    "",
+    [
+      ""
+    ],
+    "",
+    null,
+    0,
+    "",
+    0,
+    0,
+    "",
+    0,
+    "",
+    false,
+    null,
+    null,
+    [
+      null,
+      null,
+      null,
+      null,
+      "",
+      null,
+      ""
+    ],
+    null,
+    null,
+    0,
+    [],
+    0,
+    {},
+    ""
+  ],
+  [
+    "",
+    "c2",
+    "",
+    "camera",
+    [
+      [
+        "",
+        ""
+      ]
+    ],
+    "",
+    "",
+    [
+      ""
+    ],
+    "",
+    null,
+    0,
+    "",
+    0,
+    0,
+    "",
+    0,
+    "",
+    false,
+    null,
+    null,
+    [
+      null,
+      null,
+      null,
+      null,
+      "",
+      null,
+      ""
+    ],
+    null,
+    null,
+    0,
+    [],
+    0,
+    {},
+    ""
+  ],
+  [
+    "",
+    "c3",
+    "",
+    "camera",
+    [
+      [
+        "",
+        ""
+      ]
+    ],
+    "",
+    "",
+    [
+      ""
+    ],
+    "",
+    null,
+    0,
+    "",
+    0,
+    0,
+    "",
+    0,
+    "",
+    false,
+    null,
+    null,
+    [
+      null,
+      null,
+      null,
+      null,
+      "",
+      null,
+      ""
+    ],
+    null,
+    null,
+    0,
+    [],
+    0,
+    {},
+    ""
+  ],
+  [
+    "",
+    "c4",
+    "",
+    "camera",
+    [
+      [
+        "",
+        ""
+      ]
+    ],
+    "",
+    "",
+    [
+      ""
+    ],
+    "",
+    null,
+    0,
+    "",
+    0,
+    0,
+    "",
+    0,
+    "",
+    false,
+    null,
+    null,
+    [
+      null,
+      null,
+      null,
+      null,
+      "",
+      null,
+      ""
+    ],
+    null,
+    null,
+    0,
+    [],
+    0,
+    {},
+    ""
+  ],
+  [
+    "",
+    "c5",
+    "",
+    "camera",
+    [
+      [
+        "",
+        ""
+      ]
+    ],
+    "",
+    "",
+    [
+      ""
+    ],
+    "",
+    null,
+    0,
+    "",
+    0,
+    0,
+    "",
+    0,
+    "",
+    false,
+    null,
+    null,
+    [
+      null,
+      null,
+      null,
+      null,
+      "",
+      null,
+      ""
+    ],
+    null,
+    null,
+    0,
+    [],
+    0,
+    {},
+    ""
+  ],
+  [
+    "",
+    "c6",
+    "",
+    "camera",
+    [
+      [
+        "",
+        ""
+      ]
+    ],
+    "",
+    "",
+    [
+      ""
+    ],
+    "",
+    null,
+    0,
+    "",
+    0,
+    0,
+    "",
+    0,
+    "",
+    false,
+    null,
+    null,
+    [
+      null,
+      null,
+      null,
+      null,
+      "",
+      null,
+      ""
+    ],
+    null,
+    null,
+    0,
+    [],
+    0,
+    {},
+    ""
+  ],
+  [
+    "",
+    "c7",
+    "",
+    "camera",
+    [
+      [
+        "",
+        ""
+      ]
+    ],
+    "",
+    "",
+    [
+      ""
+    ],
+    "",
+    null,
+    0,
+    "",
+    0,
+    0,
+    "",
+    0,
+    "",
+    false,
+    null,
+    null,
+    [
+      null,
+      null,
+      null,
+      null,
+      "",
+      null,
+      ""
+    ],
+    null,
+    null,
+    0,
+    [],
+    0,
+    {},
+    ""
+  ],
+  [
+    "",
+    "10066761",
+    "",
+    "bridge",
+    [
+      [
+        "c0",
+        ""
+      ],
+      [
+        "c1",
+        ""
+      ],
+      [
+        "c2",
+        ""
+      ],
+      [
+        "c3",
+        ""
+      ],
+      [
+        "c4",
+        ""
+      ],
+      [
+        "c5",
+        ""
+      ],
+      [
+        "c6",
+        ""
+      ],
+      [
+        "c7",
+        ""
+      ]
+    ],
+    "",
+    "",
+    [],
+    "",
+    "",
+    0,
+    "",
+    0,
+    0,
+    "",
+    0,
+    "",
+    false,
+    null,
+    null,
+    [
+      null,
+      null,
+      null,
+      null,
+      "",
+      null,
+      ""
+    ],
+    null,
+    null,
+    0,
+    [],
+    0,
+    {},
+    ""
+  ]
+]
